Move path resolution logic to child assembly resolver class
This is necessary for the '--ref' argument to work with the mono build
of the linker. Due to the design of cecil's BaseAssemblyResolver,
there is a small amount of duplication.
