+++ src/linker/Linker/AssemblyResolver.cs
@@ -30,6 +30,7 @@ using System;
 using System.Collections.Generic;
 using System.IO;
 using Mono.Cecil;
+using Mono.Collections.Generic;
 
 namespace Mono.Linker {
 
@@ -43,6 +44,8 @@ namespace Mono.Linker {
 		HashSet<string> _unresolvedAssemblies;
 		bool _ignoreUnresolved;
 		LinkContext _context;
+		readonly Collection<string> _assemblyPaths;
+
 
 		public IDictionary<string, AssemblyDefinition> AssemblyCache {
 			get { return _assemblies; }
@@ -56,6 +59,7 @@ namespace Mono.Linker {
 		public AssemblyResolver (Dictionary<string, AssemblyDefinition> assembly_cache)
 		{
 			_assemblies = assembly_cache;
+			_assemblyPaths = new Collection<string> (10) { };
 		}
 
 		public bool IgnoreUnresolved {
@@ -68,12 +72,57 @@ namespace Mono.Linker {
 			set { _context = value; }
 		}
 
+#if !FEATURE_ILLINK
+		// The base class's definition of GetAssembly is visible when using DirectoryAssemblyResolver.
+		AssemblyDefinition GetAssembly (string file, ReaderParameters parameters)
+		{
+			if (parameters.AssemblyResolver == null)
+				parameters.AssemblyResolver = this;
+
+			return ModuleDefinition.ReadModule (file, parameters).Assembly;
+		}
+#endif
+
+		private AssemblyDefinition ResolveWithPaths (AssemblyNameReference name, ReaderParameters parameters)
+		{
+			// Validate arguments, similarly to how the base class does it.
+			if (name == null)
+				throw new ArgumentNullException ("name");
+			if (parameters == null)
+				throw new ArgumentNullException ("parameters");
+
+			AssemblyDefinition asm = null;
+
+			// Try the new resolution behavior. This can't live in the cecil-owned base class.
+			asm = SearchAssemblyPaths (name, _assemblyPaths, parameters);
+			if (asm != null)
+				return asm;
+
+			// Fall back to the base class resolution logic
+			return base.Resolve (name, parameters);
+		}
+
+		private AssemblyDefinition SearchAssemblyPaths (AssemblyNameReference name, IEnumerable<string> assemblyPaths, ReaderParameters parameters)
+		{
+			foreach (var assemblyPath in assemblyPaths) {
+				if (Path.GetFileName (assemblyPath) != name.Name + ".dll")
+					continue;
+				try {
+					return GetAssembly (assemblyPath, parameters);
+				} catch (System.BadImageFormatException) {
+					continue;
+				}
+			}
+
+			return null;
+		}
+
 		public override AssemblyDefinition Resolve (AssemblyNameReference name, ReaderParameters parameters)
 		{
 			AssemblyDefinition asm = null;
 			if (!_assemblies.TryGetValue (name.Name, out asm) && (_unresolvedAssemblies == null || !_unresolvedAssemblies.Contains (name.Name))) {
 				try {
-					asm = base.Resolve (name, parameters);
+					asm = ResolveWithPaths (name, parameters);
 					_assemblies [name.Name] = asm;
 				} catch (AssemblyResolutionException) {
 					if (!_ignoreUnresolved)
@@ -94,6 +143,11 @@ namespace Mono.Linker {
 			_assemblies [assembly.Name.Name] = assembly;
 			base.AddSearchDirectory (Path.GetDirectoryName (assembly.MainModule.FileName));
 			return assembly;
+		}
+
+		public void AddAssemblyPath (string assemblyPath)
+		{
+			_assemblyPaths.Add (assemblyPath);
 		}
 
 		protected override void Dispose (bool disposing)
