+++ src/linker/Linker/DirectoryAssemblyResolver.cs
@@ -11,16 +11,10 @@ namespace Mono.Linker {
 	public abstract class DirectoryAssemblyResolver : IAssemblyResolver {
 
 		readonly Collection<string> directories;
-		readonly Collection<string> assemblyPaths;
 
 		public void AddSearchDirectory (string directory)
 		{
 			directories.Add (directory);
-		}
-
-		public void AddAssemblyPath (string assemblyPath)
-		{
-			assemblyPaths.Add (assemblyPath);
 		}
 
 		public void RemoveSearchDirectory (string directory)
@@ -36,10 +30,9 @@ namespace Mono.Linker {
 		protected DirectoryAssemblyResolver ()
 		{
 			directories = new Collection<string> (2) { "." };
-			assemblyPaths = new Collection<string> (10) { };
 		}
 
-		AssemblyDefinition GetAssembly (string file, ReaderParameters parameters)
+		protected AssemblyDefinition GetAssembly (string file, ReaderParameters parameters)
 		{
 			if (parameters.AssemblyResolver == null)
 				parameters.AssemblyResolver = this;
@@ -57,32 +50,13 @@ namespace Mono.Linker {
 			if (name == null)
 				throw new ArgumentNullException ("name");
 			if (parameters == null)
-				parameters = new ReaderParameters ();
+				throw new ArgumentNullException ("parameters");
 
-			var assembly = SearchAssemblyPaths (name, assemblyPaths, parameters);
-			if (assembly != null)
-				return assembly;
-
-			assembly = SearchDirectory (name, directories, parameters);
+			var assembly = SearchDirectory (name, directories, parameters);
 			if (assembly != null)
 				return assembly;
 
 			throw new AssemblyResolutionException (name);
-		}
-
-		AssemblyDefinition SearchAssemblyPaths (AssemblyNameReference name, IEnumerable<string> assemblyPaths, ReaderParameters parameters)
-		{
-			foreach (var assemblyPath in assemblyPaths) {
-				if (Path.GetFileName (assemblyPath) != name.Name + ".dll")
-					continue;
-				try {
-					return GetAssembly (assemblyPath, parameters);
-				} catch (System.BadImageFormatException) {
-					continue;
-				}
-			}
-
-			return null;
 		}
 
 		AssemblyDefinition SearchDirectory (AssemblyNameReference name, IEnumerable<string> directories, ReaderParameters parameters)
